+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Added
+- add option to fix unauthorized contracts for pausing
+
 ## [0.16.0] - 2026-04-22
 ### Added
 - Global Pause flags tracked contracts where it lacks authority